Code refactoring; Added code for loading Prj2; Fixed code for saving Prj2

+++ TombEditor/Geometry/CameraInstance.cs
@@ -3,13 +3,13 @@
     public class CameraInstance : PositionBasedObjectInstance, IHasScriptID
     {
         public ushort? ScriptId { get; set; }
-        public short Sequence { get; set; }
+        public ushort Sequence { get; set; }
         public short Timer { get; set; }
-        public short Roll { get; set; }
-        public short Number { get; set; }
-        public short Speed { get; set; }
-        public short Fov { get; set; }
-        public short Flags { get; set; }
+        public float Roll { get; set; }
+        public ushort Number { get; set; }
+        public float Speed { get; set; }
+        public float Fov { get; set; }
+        public ushort Flags { get; set; }
         public bool Fixed { get; set; }
         
         public override bool CopyToFlipRooms => false;
